*** empty log message ***
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gworkspace/trunk@19178 72102866-910b-0410-8b05-ffd578937521

+++ FSNode/FSNodeRep.h
@@ -137,9 +137,9 @@ typedef enum FSNInfoType {
 
 + (float)highlightHeightFactor;
 
-+ (float)labelMargin;
++ (int)labelMargin;
 
-+ (float)defaultIconBaseShift;
++ (int)defaultIconBaseShift;
 
 + (void)setDefaultSortOrder:(int)order;
